BranchDB: Private collections and only expose objects and paths.
Colletions were never meant to be visible, they should be transparent
to the user and only be used for data structuration.

All functions available from outside the module scope are only dependent
upon objects and their paths, removing the need to have knowledge of which
collection is set as child of an object.