BranchDB: getDescendants -> getStructure
getDescendants was removed and getStructure was implemented in its
place. getStructure returns only the paths that are descendants of
a given path, it is object agnostic, and so, won't return any object
that doesn't have a path.
This speeds up the query by a factor of as much as 1000x.